+++ src/controllers/profile.controllers.ts
@@ -1,0 +1,20 @@
+import { Inject, Service } from "typedi";
+import type { Request, Response } from "express";
+import type { ProfileServices } from "../services/profile.services";
+import { errorResponse, successResponse, validationErrorResponse } from "../utils/response.utils";
+
+@Service()
+export class ProfileController {
+    constructor(@Inject("ProfileService") public profileService: ProfileServices) {}
+
+    async getMe(req: Request,res: Response){
+        try{
+        const userEmail = (req as any).user
+        const user = await this.profileService.getMe(userEmail.email)
+
+        successResponse(res, "Success Get Profile User", user, 200);
+        }catch(error){
+        errorResponse(res, "Failed login", error);
+        }
+    }
+}
